backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
requests
google-cloud-storage
PyMySQL
